Rename Controller.cs to SpaceController.cs
Also change the file header from:

'public class Controller : MonoBehaviour {'

to

'public class SpaceController : MonoBehaviour {'

The reason for this change is because the LeapMotion Core Assets package for Unity contains a nested protected class in the 'HandController' class called 'Controller' - therefore, if you are trying to use LeapMotion and SpaceNavigator in Unity, you will get a compiler error because the 'HandController' class instantiates the 'Controller' class from SpaceNavigator rather than the nested protected class. Changing the name of this 'Controller' to 'SpaceController' fixes this problem.